JAMES-3675 : Refactoring vacation mailet to be independent of JMAP
