+++ server/apps/cassandra-app/sample-configuration/mailetcontainer.xml
@@ -87,7 +87,7 @@
         </processor>
 
         <processor state="local-delivery" enableJmx="true">
-            <mailet match="All" class="org.apache.james.jmap.mailet.VacationMailet">
+            <mailet match="All" class="VacationMailet">
                 <onMailetException>ignore</onMailetException>
             </mailet>
             <mailet match="All" class="Sieve">
