+++ server/apps/distributed-app/src/test/resources/mailetcontainer.xml
@@ -94,7 +94,7 @@
 
 
         <processor state="local-delivery" enableJmx="true">
-            <mailet match="All" class="org.apache.james.jmap.mailet.VacationMailet">
+            <mailet match="All" class="VacationMailet">
                 <onMailetException>ignore</onMailetException>
             </mailet>
             <mailet match="All" class="org.apache.james.jmap.mailet.filter.JMAPFiltering">
