+++ server/container/guice/protocols/jmap/src/main/resources/defaultJmapMailetContainer.xml
@@ -46,7 +46,7 @@
             <name>bcc</name>
         </mailet>
         <mailet match="All" class="RecipientRewriteTable" />
-        <mailet match="RecipientIsLocal" class="org.apache.james.jmap.mailet.VacationMailet"/>
+        <mailet match="RecipientIsLocal" class="VacationMailet"/>
         <mailet match="RecipientIsLocal" class="org.apache.james.jmap.mailet.filter.JMAPFiltering"/>
         <mailet match="RecipientIsLocal" class="LocalDelivery"/>
         <mailet match="HostIsLocal" class="ToProcessor">
