+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/VacationReply.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.jmap.mailet;
+package org.apache.james.transport.mailets;
 
 import java.util.List;
 
@@ -25,7 +25,7 @@ import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
 
 import org.apache.james.core.MailAddress;
-import org.apache.james.jmap.draft.utils.MimeMessageBodyGenerator;
+import org.apache.james.transport.util.MimeMessageBodyGenerator;
 import org.apache.james.vacation.api.Vacation;
 import org.apache.mailet.Mail;
 import org.apache.mailet.base.AutomaticallySentMailDetector;
