+++ server/mailet/mailets/src/test/resources/noReplyTo.eml
@@ -1,0 +1,69 @@
+Return-Path: <root-bounces@listes.apache.org>
+Content-Type: multipart/mixed; boundary="----------=_1433322346-12583-0"
+MIME-Version: 1.0
+From: "Content-filter at spam.apache.org" <postmaster@apache.org>
+Date: Wed, 3 Jun 2015 09:05:46 +0000 (UTC)
+To: <root@listes.apache.org>
+Message-ID: <VASs-IZaXqmZao@spam.apache.org>
+Subject: Original subject
+X-BeenThere: root@listes.apache.org
+Errors-To: root-bounces@listes.apache.org
+Sender: "root" <root-bounces@listes.apache.org>
+Reply-To: <>
+
+This is a multi-part message in MIME format...
+
+------------=_1433322346-12583-0
+Content-Type: text/plain; charset="UTF-8"
+Content-Disposition: inline
+Content-Transfer-Encoding: 7bit
+
+No viruses were found.
+
+Content type: Unchecked
+Internal reference code for the message is 12583-16/Ss-IZaXqmZao
+
+According to a 'Received:' trace, the message apparently originated at:
+  [198.252.153.129], [127.0.0.1] localhost [127.0.0.1] Authenticated sender:
+  quentin.h
+
+Return-Path: <quentin.h@apache.org>
+From: Quentin <quentin.h@apache.org>
+Message-ID: <556EC361.6000308@apache.org>
+Subject: XYZ
+Not quarantined.
+
+The message WILL BE relayed to:
+<xxx@apache.org>
+
+
+------------=_1433322346-12583-0
+Content-Type: text/rfc822-headers; name="header"
+Content-Disposition: inline; filename="header"
+Content-Transfer-Encoding: 7bit
+Content-Description: Message header section
+
+Return-Path: <quentin.h@apache.org>
+Message-ID: <556EC361.6000308@apache.org>
+Date: Wed, 03 Jun 2015 11:05:37 +0200
+From: Quentin <quentin.h@apache.org>
+MIME-Version: 1.0
+To: Yann <yann@apache.org>
+Subject: XYZ
+References: <556E332B.8020808@wootdevices.io> <556E3512.3050404@apache.org>
+In-Reply-To: <556E3512.3050404@apache.org>
+Content-Type: multipart/encrypted;
+ protocol="application/pgp-encrypted";
+ boundary="M0xVhKIvXqi85dG57o5RfCUAoFwhAw1Nh"
+
+------------=_1433322346-12583-0
+Content-Type: text/plain; charset="iso-8859-1"
+MIME-Version: 1.0
+Content-Transfer-Encoding: quoted-printable
+Content-Disposition: inline
+
+_______________________________________________
+root mailing list
+root@listes.apache.org
+https://listes.apache.org/cgi-bin/mailman/listinfo/root
+------------=_1433322346-12583-0--
