+++ server/mailet/mailets/src/test/resources/spamMail.eml
@@ -1,0 +1,135 @@
+Return-Path: <root-bounces@listes.minet.net>
+Received: from mx1.minet.net (mx1.minet.net [192.168.102.25])
+	 by imap (Cyrus v2.4.16-Debian-2.4.16-4+deb7u1) with LMTPA;
+	 Wed, 03 Jun 2015 11:05:45 +0200
+X-Sieve: CMU Sieve 2.4
+Received: from smtp.minet.net (smtp.minet.net [192.168.102.18])
+	by mx1.minet.net (Postfix) with ESMTP id F06FE610F72;
+	Wed,  3 Jun 2015 11:05:48 +0200 (CEST)
+Received: from listes.minet.net (listes.minet.net [192.168.102.29])
+	by smtp.minet.net (Postfix) with ESMTP id AAB15610F70;
+	Wed,  3 Jun 2015 09:05:48 +0000 (UTC)
+DKIM-Signature: v=1; a=rsa-sha256; c=simple/simple; d=minet.net; s=mail;
+	t=1433322348; bh=8OtmVRMX7IWWyQo8WQIi5rKeDEQFghSX2r1WN9/pqUk=;
+	h=From:Date:To:Subject:List-Id:List-Unsubscribe:List-Post:List-Help:
+	 List-Subscribe:From;
+	b=UkthSFXn5mtPnn7sYvDJIxgCk1O36L0p+hTOnI/xTF2WRMGJyXtLyCdFdE5jQ5gE4
+	 EGNdhwmfjaKsoE7IDtc/4GrPvbnM8i9ojpSOBW1hXJPOeYnRgah/4gsXsMJWvmmq/A
+	 4oTnw8ZswFXiShDmBMyaSEZuVrWTdaPBgUhVhGw8=
+Received: from listes.minet.net (listes.minet.net [127.0.0.1])
+	by listes.minet.net (Postfix) with ESMTP id 9CD5BA27FF9;
+	Wed,  3 Jun 2015 09:05:48 +0000 (UTC)
+X-Original-To: root@listes.minet.net
+Delivered-To: root@listes.minet.net
+Received: from mx1.minet.net (mx1.minet.net [192.168.102.25])
+ by listes.minet.net (Postfix) with ESMTP id B8C1FA27FEF
+ for <root@listes.minet.net>; Wed,  3 Jun 2015 09:05:46 +0000 (UTC)
+Received: from localhost (spam.minet.net [192.168.102.97])
+ by mx1.minet.net (Postfix) with ESMTP id AD28B610F6D
+ for <root@listes.minet.net>; Wed,  3 Jun 2015 11:05:46 +0200 (CEST)
+Content-Type: multipart/mixed; boundary="----------=_1433322346-12583-0"
+Content-Transfer-Encoding: 7bit
+MIME-Version: 1.0
+From: "Content-filter at spam.minet.net" <postmaster@minet.net>
+Date: Wed, 3 Jun 2015 09:05:46 +0000 (UTC)
+To: <root@listes.minet.net>
+Message-ID: <VASs-IZaXqmZao@spam.minet.net>
+Subject: Original subject
+X-BeenThere: root@listes.minet.net
+X-Mailman-Version: 2.1.15
+Errors-To: root-bounces@listes.minet.net
+Sender: "root" <root-bounces@listes.minet.net>
+
+This is a multi-part message in MIME format...
+
+------------=_1433322346-12583-0
+Content-Type: text/plain; charset="UTF-8"
+Content-Disposition: inline
+Content-Transfer-Encoding: 7bit
+
+No viruses were found.
+
+Content type: Unchecked
+Internal reference code for the message is 12583-16/Ss-IZaXqmZao
+
+According to a 'Received:' trace, the message apparently originated at:
+  [198.252.153.129], [127.0.0.1] localhost [127.0.0.1] Authenticated sender:
+  quentin.h
+
+Return-Path: <quentin.h@riseup.net>
+From: Quentin <quentin.h@riseup.net>
+Message-ID: <556EC361.6000308@riseup.net>
+Subject: =?UTF-8?B?UmU6IEZ3ZDogW1RtcGxhYl0gW0FQUEVMIEEgUEFSVElDSVBBVElPTl0=?=
+  =?UTF-8?B?IFdvb3QgZGV2aWNlcyAjMiAgOiB2ZW5leiBmYWlyZSBsZXVyIGhhY2tmw6p0ZSA=?=
+  =?UTF-8?B?YXV4IHBldGl0ZXMgbWFjaGluZXMgcHJvZ3JhbW1hYmxlcyBsZXMgNi03IGp1aW4=?=
+  =?UTF-8?B?IGF1IEphcmRpbiBkJ2FsaWNlL0JsYWNrbG9vcA==?=
+Not quarantined.
+
+The message WILL BE relayed to:
+<yann@minet.net>
+
+
+------------=_1433322346-12583-0
+Content-Type: text/rfc822-headers; name="header"
+Content-Disposition: inline; filename="header"
+Content-Transfer-Encoding: 7bit
+Content-Description: Message header section
+
+Return-Path: <quentin.h@riseup.net>
+Received-SPF: Pass (sender SPF authorized) identity=mailfrom; client-ip=198.252.153.129; helo=mx1.riseup.net; envelope-from=quentin.h@riseup.net; receiver=yann@minet.net
+Authentication-Results: mx2.minet.net; dkim=pass
+	reason="1024-bit key; insecure key"
+	header.d=riseup.net header.i=@riseup.net header.b=RVs2xP3R;
+	dkim-adsp=pass; dkim-atps=neutral
+Received: from mx1.riseup.net (mx1.riseup.net [198.252.153.129])
+	(using TLSv1.2 with cipher ECDHE-RSA-AES256-GCM-SHA384 (256/256 bits))
+	(No client certificate requested)
+	by mx2.minet.net (Postfix) with ESMTPS id 6D9E0A27FCA
+	for <yann@minet.net>; Wed,  3 Jun 2015 11:05:44 +0200 (CEST)
+Received: from plantcutter.riseup.net (plantcutter-pn.riseup.net [10.0.1.121])
+	(using TLSv1 with cipher ECDHE-RSA-AES256-SHA (256/256 bits))
+	(Client CN "*.riseup.net", Issuer "COMODO RSA Domain Validation Secure Server CA" (verified OK))
+	by mx1.riseup.net (Postfix) with ESMTPS id 5CC6541A75;
+	Wed,  3 Jun 2015 09:05:42 +0000 (UTC)
+DKIM-Signature: v=1; a=rsa-sha256; c=relaxed/simple; d=riseup.net; s=squak;
+	t=1433322342; bh=Lz3x5ntYzOhQOpyviKamr5VE0TinfpBCnsivvJzojYk=;
+	h=Date:From:To:Subject:References:In-Reply-To:From;
+	b=RVs2xP3RV4zWi3Jbs6p85cO8s44yLq1LXxwe6NQvgZjsHQdWaMY16jQzSbCRRdCRY
+	 IesTpg1mncqIinhXxL9grw2+0MgO+E0HM0ZPZzXM/xNTGZpALI1X6tDM0rxB+GEnHa
+	 NfGRbWnesWFLyba7am02KbTKeKq7svIBBmSbaFsE=
+Received: from [127.0.0.1] (localhost [127.0.0.1])
+	(Authenticated sender: quentin.h)
+	with ESMTPSA id 6DBAC20827
+Message-ID: <556EC361.6000308@riseup.net>
+Date: Wed, 03 Jun 2015 11:05:37 +0200
+From: Quentin <quentin.h@riseup.net>
+User-Agent: Mozilla/5.0 (X11; Linux x86_64; rv:31.0) Gecko/20100101 Thunderbird/31.7.0
+MIME-Version: 1.0
+To: Yann Sionneau <yann@minet.net>, =?UTF-8?B?U3TDqXBoYW5pZSBPdWlsbG9u?=
+ <stephanie@ouillon.fr>, =?UTF-8?B?TmF0YWxpYSBDYWxkZXJvbiBCZWx0csOhbg==?=
+ <natalia@calderon.be>
+Subject: =?UTF-8?B?UmU6IEZ3ZDogW1RtcGxhYl0gW0FQUEVMIEEgUEFSVElDSVBBVElPTl0=?=
+ =?UTF-8?B?IFdvb3QgZGV2aWNlcyAjMiAgOiB2ZW5leiBmYWlyZSBsZXVyIGhhY2tmw6p0ZSA=?=
+ =?UTF-8?B?YXV4IHBldGl0ZXMgbWFjaGluZXMgcHJvZ3JhbW1hYmxlcyBsZXMgNi03IGp1aW4=?=
+ =?UTF-8?B?IGF1IEphcmRpbiBkJ2FsaWNlL0JsYWNrbG9vcA==?=
+References: <556E332B.8020808@wootdevices.io> <556E3512.3050404@minet.net>
+In-Reply-To: <556E3512.3050404@minet.net>
+OpenPGP: id=AA5EE4B4EA20E2BEAA70B98E49CBF006922B990A;
+	url=pgp.mit.edu
+Content-Type: multipart/encrypted;
+ protocol="application/pgp-encrypted";
+ boundary="M0xVhKIvXqi85dG57o5RfCUAoFwhAw1Nh"
+X-Virus-Scanned: clamav-milter 0.98.7 at mx1
+X-Virus-Status: Clean
+
+------------=_1433322346-12583-0
+Content-Type: text/plain; charset="iso-8859-1"
+MIME-Version: 1.0
+Content-Transfer-Encoding: quoted-printable
+Content-Disposition: inline
+
+_______________________________________________
+root mailing list
+root@listes.minet.net
+https://listes.minet.net/cgi-bin/mailman/listinfo/root
+------------=_1433322346-12583-0--
